+++ lib/sgtn-client/core/cache.rb
@@ -6,6 +6,7 @@ module SgtnClient::Core
 
             def self.initialize(disabled=false, opts={})
                 $opts = opts
+                @mutex = Mutex.new
                 SgtnClient.logger.debug "Initialize cache......"
                 if disabled == false
                     $data = Hash.new
@@ -41,42 +42,50 @@ module SgtnClient::Core
             end
 
             def self.put(key, value, ttl=nil)
-                if $data == nil
-                    return nil
+                @mutex.synchronize do
+                    if $data == nil
+                        return nil
+                    end
+                    ttl ||= @opts[:ttl]
+                    # hours from new
+                    SgtnClient.logger.debug "Put cache for key '" + key + "' with expired time at'" + (Time.now + ttl*60).to_s
+                    $data[key] = Entry.new(Time.now + ttl*60, value)
                 end
-                ttl ||= @opts[:ttl]
-                # hours from new
-                SgtnClient.logger.debug "Put cache for key '" + key + "' with expired time at'" + (Time.now + ttl*60).to_s
-                $data[key] = Entry.new(Time.now + ttl*60, value)
             end
 
             def self.delete(key)
-                if $data == nil
-                    return nil
+                @mutex.synchronize do
+                    if $data == nil
+                        return nil
+                    end
+                    SgtnClient.logger.debug "Delete cache for key: " + key
+                    $data.delete key
                 end
-                SgtnClient.logger.debug "Delete cache for key: " + key
-                $data.delete key
             end
 
             def self.clear
-                if $data == nil
-                    return nil
+                @mutex.synchronize do
+                    if $data == nil
+                        return nil
+                    end
+                    SgtnClient.logger.debug "Clear cache!"
+                    $data = Hash.new
                 end
-                SgtnClient.logger.debug "Clear cache!"
-                $data = Hash.new
             end
 
             def self.invalidate
-                if $data == nil
-                    return nil
+                @mutex.synchronize do
+                    if $data == nil
+                        return nil
+                    end
+                    SgtnClient.logger.debug "Invalidating expired cache......"
+                    now = Time.now
+                    $data.each {
+                        |k, v|
+                            SgtnClient.logger.debug "Checking cache: key=#{k}, expiredtime=#{v[:expiry]}, now=#{now}, expired=#{(v[:expiry] < now)}"
+                        }
+                    $data.delete_if {|k, v| v[:expiry] < now}
                 end
-                SgtnClient.logger.debug "Invalidating expired cache......"
-                now = Time.now
-                $data.each {
-                    |k, v|
-                        SgtnClient.logger.debug "Checking cache: key=#{k}, expiredtime=#{v[:expiry]}, now=#{now}, expired=#{(v[:expiry] < now)}"
-                    }
-                $data.delete_if {|k, v| v[:expiry] < now}
             end
         end
