+++ lib/sgtn-client/sgtn-client.rb
@@ -11,6 +11,7 @@ module SgtnClient
       autoload :Exceptions,         "sgtn-client/core/exceptions"
       autoload :ValidateUtil,       "sgtn-client/util/validate-util"
       autoload :LocaleUtil,         "sgtn-client/util/locale-util"
+      autoload :FileUtil,           "sgtn-client/util/file-util"
 
       module Formatters
             autoload :PluralFormatter,    "sgtn-client/formatters/plurals/plural_formatter"
